# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nombre de tu API",
        "description": "Descripción de tu API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/exportAPI": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Export users data to Excel",
                "description": "Allows exporting users data to Excel format",
                "operationId": "exportUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/importExcelDataAPI": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Import users data from Excel",
                "description": "Allows importing users data from an Excel file",
                "operationId": "importUsers",
                "requestBody": {
                    "description": "Excel file to import",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "import_file"
                                ],
                                "properties": {
                                    "import_file": {
                                        "description": "Excel file to import",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Importación exitosa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    }
}